User controller api

See merge request s09-webmobile1-sub2/S09P12B303!29

+++ Server/src/main/java/com/threeracha/gaewoonhae/db/domain/UserBuyEmoji.java
@@ -17,10 +17,10 @@ public class UserBuyEmoji {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "buy_id")
-    private int buyId;
+    private Long buyId;
 
     @Column(name = "user_id", nullable = false)
-    private int userId;
+    private Long userId;
 
     @Column(name = "emoji_id", nullable = false)
     private int emojiId;
